* UiaAtkBridge/UiaAtkBridge/AmbiDictionary.cs,   UiaAtkBridge/UiaAtkBridge/Monitor.cs: Cosmetic. * UiaAtkBridge/UiaAtkBridge/Adapter.cs,   UiaAtkBridge/UiaAtkBridge/Button.cs: Track API changes. * UiaAtkBridge/UiaAtkBridge/AutomationBridge.cs,   UIAutomationWinforms/UIAutomationWinforms/Mono.UIAutomation.Winforms/FormListener.cs:   Report all form removals (seems to fix bug 427857)
svn path=/trunk/uia2atk/; revision=114533

+++ src/UiaAtkBridge/UiaAtkBridge/Adapter.cs
@@ -51,16 +51,16 @@ namespace UiaAtkBridge
 		{
 			if (e.Property == AutomationElementIdentifiers.HasKeyboardFocusProperty) {
 				bool focused = (bool)e.NewValue;
-				NotifyStateChange ((ulong) Atk.StateType.Focused, focused);
+				NotifyStateChange (Atk.StateType.Focused, focused);
 				if (focused)
 					Atk.Focus.TrackerNotify (this);
 			} else if (e.Property == AutomationElementIdentifiers.IsOffscreenProperty) { 
 				bool offscreen = (bool)e.NewValue;
-				NotifyStateChange ((ulong) Atk.StateType.Visible, !offscreen);
+				NotifyStateChange (Atk.StateType.Visible, !offscreen);
 			} else if (e.Property == AutomationElementIdentifiers.IsEnabledProperty) {
 				bool enabled = (bool) e.NewValue;
-				NotifyStateChange ((ulong) Atk.StateType.Enabled, enabled);
-				NotifyStateChange ((ulong) Atk.StateType.Sensitive, enabled);
+				NotifyStateChange (Atk.StateType.Enabled, enabled);
+				NotifyStateChange (Atk.StateType.Sensitive, enabled);
 			} else if (e.Property == AutomationElementIdentifiers.NameProperty) {
 				Name = (string)e.NewValue;
 			}
